Added additional wazuh-dashboard checks
Added code to check that the wazuh dashboard service exists as well as the /etc/wazuh-dashboard directory. This is an attempt to solve an issue I ran into with the installation assistant not properly installing the dashboard or certificates by alerting the user if the aforementioned items do not exist.